chris build out
