+++ lib/site_builder.rb
@@ -2,39 +2,7 @@ require 'fileutils'
 require 'kramdown'
 require_relative 'file_reader'
 require_relative 'file_writer'
-
-
-class FileManager
-
-  def create_new_site_structure(filename)
-    if Dir.exist?(filename)
-      result =  "Error. Directory #{filename} already exists."
-    else
-      Dir.mkdir(filename)
-      Dir.mkdir(filename + "/_output")
-      Dir.mkdir(filename + "/source")
-      make_sub_folders(filename + "/source")
-      make_starting_files(filename + "/source")
-      result = "File structure created"
-    end
-    result
-  end
-
-  def make_sub_folders(filename)
-    Dir.mkdir(filename + "/css")
-    Dir.mkdir(filename + "/pages")
-    Dir.mkdir(filename + "/posts")
-    Dir.mkdir(filename + "/layouts")
-  end
-
-  def make_starting_files(filename)
-    File.write(filename + "/css/main.css", "")
-    File.write(filename + "/index.markdown", "")
-    File.write(filename + "/pages/about.markdown", "")
-    t = Time.new
-    File.write(filename + "/posts/" + t.strftime("%F") + "-welcome-to-hyde.markdown", "")
-  end
-end
+require_relative 'file_manager'
 
 class SiteBuilder
 
@@ -83,6 +51,7 @@ class SiteBuilder
 
   end
 
+
 # Upon creation, markdown should pre-poulate some basic markdown i.e.
 #Blog Name
 ## Writen on date
@@ -110,6 +79,8 @@ if __FILE__ ==$0
     when nil
       puts "No command given"
     when "new"
+        # filename = gets
+        # filename.chomp!
       f=FileManager.new
       f.create_new_site_structure(filename)
     when "build"
@@ -118,35 +89,3 @@ if __FILE__ ==$0
     end
   end
 end
-
-
-
-
-
-  # if ARGV[0] == nil
-  #   puts "No command given"
-  # elsif ARGV[0] == "new"
-  #
-  #
-  #
-  # end
-  # end
-
-
-
-# args = ARGV
-# puts "You ran Hyde with these arguments: #{args}"
-
-# ##############
-# Your code here
-# ##############
-
-#
-# class Hyde
-#  def initialize(method, location)
-#   Dir.mkdir("method")
-#   p "#{method}#{location}"
-#  end
-#
-#
-# end
